+++ src/app/user/[userId]/_components/Card.tsx
@@ -1,15 +1,12 @@
-import { MouseEvent } from 'react';
 import { assignInlineVars } from '@vanilla-extract/dynamic';
 
 import * as styles from './Card.css';
 
 import useMoveToPage from '@/hooks/useMoveToPage';
-import useBooleanOutput from '@/hooks/useBooleanOutput';
-
-import PopupMenuIcon from '/public/icons/popup_menu.svg';
 import { ListType } from '@/lib/types/listType';
 import { BACKGROUND_COLOR_READ } from '@/styles/Color';
-import PopupMenu from './PopupMenu';
+
+import OptionToggleButton from './OptionToggleButton';
 
 interface CardProps {
   list: ListType;
@@ -17,14 +14,16 @@ interface CardProps {
   userId: number;
 }
 
+/**
+TODO
+- [x] 팝오버 메뉴 토글 기능
+- [x] 외부 클릭 시 팝오버 닫히는 기능
+- [ ] 팝오버 메뉴에 삭제하기 추가
+- [ ] 리스트 삭제하기 기능
+ */
+
 export default function Card({ list, isOwner }: CardProps) {
   const { onClickMoveToPage } = useMoveToPage();
-  const { isOn: popupIsOpen, toggle: popupToggle, handleSetOn: handlePopupOn } = useBooleanOutput();
-
-  const handleOpenMenu = (e: MouseEvent<HTMLButtonElement>) => {
-    e.stopPropagation();
-    handlePopupOn();
-  };
 
   return (
     // MasonryGrid 라이브러리에서는 ul로 감싸줘야 하므로 Link태그 미사용
@@ -38,10 +37,7 @@ export default function Card({ list, isOwner }: CardProps) {
       {isOwner && (
         <div className={styles.label}>
           <div className={styles.labelText}>{list.isPublic ? '공개' : '비공개'}</div>
-          <button className={styles.labelOption} onClick={handleOpenMenu}>
-            <PopupMenuIcon alt="리스트 공개, 비공개 옵션" />
-          </button>
-          {popupIsOpen && <PopupMenu />}
+          <OptionToggleButton />
         </div>
       )}
       <h2 className={styles.title}>{list.title}</h2>
